Reorder rhsmlib tests
Until now, the tests in 'rhsmlib_test/' directory were testing two
things at once: services (which provide the functionality to CLI and
DBus) and DBus (which implements the com.redhat.RHSM1 API).

The files have not been altered in any way. Imports were fixed to match
new paths and unused imports have been removed.

(Adapted from 3a2a5fd)
