+++ test/rhsmlib_test/facts/dmidecodedata/aarch64-baremetal-1.txt
