+++ test/rhsmlib_test/facts/dmidecodedata/aarch64-baremetal-2.txt
