+++ test/rhsmlib_test/facts/dmidecodedata/aarch64-qemu-1.txt
