+++ test/rhsmlib_test/facts/dmidecodedata/x86_64-aws-1.txt
