+++ test/rhsmlib_test/facts/dmidecodedata/x86_64-baremetal-1.txt
