+++ test/rhsmlib_test/facts/dmidecodedata/x86_64-esx6.7-1.txt
