+++ test/rhsmlib_test/facts/dmidecodedata/x86_64-qemu-1.txt
