+++ test/rhsmlib_test/facts/test_virt.py
