+++ test/rhsmlib_test/services/test_register.py
@@ -10,14 +10,9 @@
 # Red Hat trademarks are not licensed under GPLv2. No permission is
 # granted to use or replicate Red Hat trademarks that are incorporated
 # in this software or its documentation.
-#
-
-import errno
 
 import mock
 import json
-import dbus.connection
-import socket
 
 import subscription_manager.injection as inj
 
@@ -27,20 +22,15 @@ from subscription_manager.facts import Facts
 from subscription_manager.identity import Identity
 from subscription_manager.plugins import PluginManager
 
-from ..fixture import set_up_mock_sp_store
+from test.fixture import set_up_mock_sp_store
 
-from test.rhsmlib_test.base import DBusObjectTest, InjectionMockingTest
+from test.rhsmlib_test.base import InjectionMockingTest
 
 from rhsm import connection
 
-from rhsmlib.dbus import constants
-from rhsmlib.dbus.objects import RegisterDBusObject
-
 from rhsmlib.services import register, exceptions
 
-from test import subman_marker_dbus
-
-CONSUMER_CONTENT_JSON = '''{"hypervisorId": null,
+CONSUMER_CONTENT_JSON = """{"hypervisorId": null,
         "serviceLevel": "",
         "autoheal": true,
         "idCert": {
@@ -77,10 +67,10 @@ CONSUMER_CONTENT_JSON = '''{"hypervisorId": null,
         "lastCheckin": null, "entitlementCount": 0, "releaseVer":
         {"releaseVer": null}, "entitlementStatus": "valid", "name":
         "test.example.com", "created": "2016-06-02T15:16:51+0000",
-        "contentTags": null, "dev": false}'''
+        "contentTags": null, "dev": false}"""
 
 # Following consumer do not contain information about content access mode
-OLD_CONSUMER_CONTENT_JSON = '''{"hypervisorId": null,
+OLD_CONSUMER_CONTENT_JSON = """{"hypervisorId": null,
         "serviceLevel": "",
         "autoheal": true,
         "idCert": {
@@ -116,9 +106,9 @@ OLD_CONSUMER_CONTENT_JSON = '''{"hypervisorId": null,
         "lastCheckin": null, "entitlementCount": 0, "releaseVer":
         {"releaseVer": null}, "entitlementStatus": "valid", "name":
         "test.example.com", "created": "2016-06-02T15:16:51+0000",
-        "contentTags": null, "dev": false}'''
+        "contentTags": null, "dev": false}"""
 
-OWNERS_CONTENT_JSON = '''[
+OWNERS_CONTENT_JSON = """[
     {
         "autobindDisabled": false,
         "autobindHypervisorDisabled": false,
@@ -174,7 +164,7 @@ OWNERS_CONTENT_JSON = '''[
         "upstreamConsumer": null
     }
 ]
-'''
+"""
 
 
 class RegisterServiceTest(InjectionMockingTest):
@@ -192,22 +182,22 @@ class RegisterServiceTest(InjectionMockingTest):
 
         # Add a mock for content access mode cache
         self.mock_content_access_mode_cache = mock.Mock(
-            spec=ContentAccessModeCache,
-            name="ContentAccessModeCache"
+            spec=ContentAccessModeCache, name="ContentAccessModeCache"
         )
-        self.mock_content_access_mode_cache.read_data = mock.Mock(return_value='entitlement')
+        self.mock_content_access_mode_cache.read_data = mock.Mock(return_value="entitlement")
 
         # For the tests in which it's used, the consumer_auth cp and basic_auth cp can be the same
         self.mock_cp_provider.get_consumer_auth_cp.return_value = self.mock_cp
         self.mock_cp_provider.get_basic_auth_cp.return_value = self.mock_cp
 
         self.mock_pm = mock.Mock(spec=PluginManager, name="PluginManager")
-        self.mock_installed_products = mock.Mock(spec=InstalledProductsManager,
-                                                 name="InstalledProductsManager")
+        self.mock_installed_products = mock.Mock(
+            spec=InstalledProductsManager, name="InstalledProductsManager"
+        )
         self.mock_facts = mock.Mock(spec=Facts, name="Facts")
         self.mock_facts.get_facts.return_value = {}
 
-        syspurpose_patch = mock.patch('subscription_manager.syspurposelib.SyncedStore')
+        syspurpose_patch = mock.patch("subscription_manager.syspurposelib.SyncedStore")
         self.mock_sp_store = syspurpose_patch.start()
         self.mock_sp_store, self.mock_sp_store_contents = set_up_mock_sp_store(self.mock_sp_store)
         self.addCleanup(syspurpose_patch.stop)
@@ -255,14 +245,15 @@ class RegisterServiceTest(InjectionMockingTest):
             role="",
             addons=[],
             service_level="",
-            usage="")
+            usage="",
+        )
         self.mock_installed_products.write_cache.assert_called()
 
         mock_persist_consumer.assert_called_once_with(expected_consumer)
         mock_write_cache.assert_called_once()
         expected_plugin_calls = [
-            mock.call('pre_register_consumer', name='name', facts={}),
-            mock.call('post_register_consumer', consumer=expected_consumer, facts={})
+            mock.call("pre_register_consumer", name="name", facts={}),
+            mock.call("post_register_consumer", consumer=expected_consumer, facts={}),
         ]
         self.assertEqual(expected_plugin_calls, self.mock_pm.run.call_args_list)
 
@@ -295,19 +286,20 @@ class RegisterServiceTest(InjectionMockingTest):
             role="",
             addons=[],
             service_level="",
-            usage="")
+            usage="",
+        )
         self.mock_installed_products.write_cache.assert_called()
 
         mock_persist_consumer.assert_called_once_with(expected_consumer)
         mock_write_cache.assert_called_once()
         expected_plugin_calls = [
-            mock.call('pre_register_consumer', name='name', facts={}),
-            mock.call('post_register_consumer', consumer=expected_consumer, facts={})
+            mock.call("pre_register_consumer", name="name", facts={}),
+            mock.call("post_register_consumer", consumer=expected_consumer, facts={}),
         ]
         self.assertEqual(expected_plugin_calls, self.mock_pm.run.call_args_list)
-        assert 'owner' in consumer
-        assert 'contentAccessMode' in consumer['owner']
-        assert 'entitlement' == consumer['owner']['contentAccessMode']
+        assert "owner" in consumer
+        assert "contentAccessMode" in consumer["owner"]
+        assert "entitlement" == consumer["owner"]["contentAccessMode"]
 
     @mock.patch("rhsmlib.services.register.syspurposelib.write_syspurpose_cache", return_value=True)
     @mock.patch("rhsmlib.services.register.managerlib.persist_consumer_cert")
@@ -319,7 +311,7 @@ class RegisterServiceTest(InjectionMockingTest):
         self.mock_installed_products.format_for_server.return_value = []
         self.mock_installed_products.tags = []
         expected_consumer = json.loads(OLD_CONSUMER_CONTENT_JSON)
-        del expected_consumer['owner']
+        del expected_consumer["owner"]
         self.mock_cp.registerConsumer.return_value = expected_consumer
 
         register_service = register.RegisterService(self.mock_cp)
@@ -338,17 +330,18 @@ class RegisterServiceTest(InjectionMockingTest):
             role="",
             addons=[],
             service_level="",
-            usage="")
+            usage="",
+        )
         self.mock_installed_products.write_cache.assert_called()
 
         mock_persist_consumer.assert_called_once_with(expected_consumer)
         mock_write_cache.assert_called_once()
         expected_plugin_calls = [
-            mock.call('pre_register_consumer', name='name', facts={}),
-            mock.call('post_register_consumer', consumer=expected_consumer, facts={})
+            mock.call("pre_register_consumer", name="name", facts={}),
+            mock.call("post_register_consumer", consumer=expected_consumer, facts={}),
         ]
         self.assertEqual(expected_plugin_calls, self.mock_pm.run.call_args_list)
-        assert 'owner' not in consumer
+        assert "owner" not in consumer
 
     @mock.patch("rhsmlib.services.register.syspurposelib.write_syspurpose_cache", return_value=True)
     @mock.patch("rhsmlib.services.register.managerlib.persist_consumer_cert")
@@ -375,7 +368,7 @@ class RegisterServiceTest(InjectionMockingTest):
                 "contentAccessMode": "entitlement",
                 "contentAccessModeList": "entitlement,org_environment",
                 "lastRefreshed": None,
-                "href": "/owners/snowwhite"
+                "href": "/owners/snowwhite",
             }
         ]
         self.mock_identity.is_valid.return_value = False
@@ -393,9 +386,7 @@ class RegisterServiceTest(InjectionMockingTest):
             pass
 
         org = register_service.determine_owner_key(
-            username=self.mock_cp.username,
-            get_owner_cb=_get_owner_cb,
-            no_owner_cb=_no_owner_cb
+            username=self.mock_cp.username, get_owner_cb=_get_owner_cb, no_owner_cb=_no_owner_cb
         )
 
         self.assertIsNotNone(org)
@@ -415,14 +406,15 @@ class RegisterServiceTest(InjectionMockingTest):
             role="",
             addons=[],
             service_level="",
-            usage="")
+            usage="",
+        )
         self.mock_installed_products.write_cache.assert_called()
 
         mock_persist_consumer.assert_called_once_with(expected_consumer)
         mock_write_cache.assert_called_once()
         expected_plugin_calls = [
-            mock.call('pre_register_consumer', name='name', facts={}),
-            mock.call('post_register_consumer', consumer=expected_consumer, facts={})
+            mock.call("pre_register_consumer", name="name", facts={}),
+            mock.call("post_register_consumer", consumer=expected_consumer, facts={}),
         ]
         self.assertEqual(expected_plugin_calls, self.mock_pm.run.call_args_list)
 
@@ -454,15 +446,15 @@ class RegisterServiceTest(InjectionMockingTest):
             role="",
             addons=[],
             service_level="",
-            usage=""
+            usage="",
         )
         self.mock_installed_products.write_cache.assert_called()
 
         mock_persist_consumer.assert_called_once_with(expected_consumer)
         mock_write_cache.assert_called_once()
         expected_plugin_calls = [
-            mock.call('pre_register_consumer', name='name', facts={}),
-            mock.call('post_register_consumer', consumer=expected_consumer, facts={})
+            mock.call("pre_register_consumer", name="name", facts={}),
+            mock.call("post_register_consumer", consumer=expected_consumer, facts={}),
         ]
         self.assertEqual(expected_plugin_calls, self.mock_pm.run.call_args_list)
 
@@ -485,24 +477,24 @@ class RegisterServiceTest(InjectionMockingTest):
         mock_persist_consumer.assert_called_once_with(expected_consumer)
         mock_write_cache.assert_called_once()
         expected_plugin_calls = [
-            mock.call('pre_register_consumer', name='name', facts={}),
-            mock.call('post_register_consumer', consumer=expected_consumer, facts={})
+            mock.call("pre_register_consumer", name="name", facts={}),
+            mock.call("post_register_consumer", consumer=expected_consumer, facts={}),
         ]
         self.assertEqual(expected_plugin_calls, self.mock_pm.run.call_args_list)
 
     def _build_options(self, activation_keys=None, environments=None, force=None, name=None, consumerid=None):
         return {
-            'activation_keys': activation_keys,
-            'environments': environments,
-            'force': force,
-            'name': name,
-            'consumerid': consumerid
+            "activation_keys": activation_keys,
+            "environments": environments,
+            "force": force,
+            "name": name,
+            "consumerid": consumerid,
         }
 
     def test_fails_when_previously_registered(self):
         self.mock_identity.is_valid.return_value = True
 
-        with self.assertRaisesRegex(exceptions.ValidationError, r'.*system is already registered.*'):
+        with self.assertRaisesRegex(exceptions.ValidationError, r".*system is already registered.*"):
             register.RegisterService(self.mock_cp).validate_options(self._build_options())
 
     def test_allows_force(self):
@@ -540,7 +532,8 @@ class RegisterServiceTest(InjectionMockingTest):
             role="test_role",
             service_level="test_sla",
             type="system",
-            usage="test_usage")
+            usage="test_usage",
+        )
         mock_write_cache.assert_called_once()
 
     def test_does_not_require_basic_auth_with_activation_keys(self):
@@ -554,7 +547,7 @@ class RegisterServiceTest(InjectionMockingTest):
     def test_does_not_allow_basic_auth_with_activation_keys(self):
         self.mock_identity.is_valid.return_value = False
         options = self._build_options(activation_keys=[1])
-        with self.assertRaisesRegex(exceptions.ValidationError, r'.*do not require user credentials.*'):
+        with self.assertRaisesRegex(exceptions.ValidationError, r".*do not require user credentials.*"):
             register.RegisterService(self.mock_cp).validate_options(options)
 
     def test_does_not_allow_environment_with_activation_keys(self):
@@ -562,8 +555,8 @@ class RegisterServiceTest(InjectionMockingTest):
         self.mock_cp.password = None
 
         self.mock_identity.is_valid.return_value = False
-        options = self._build_options(activation_keys=[1], environments='environment')
-        with self.assertRaisesRegex(exceptions.ValidationError, r'.*do not allow environments.*'):
+        options = self._build_options(activation_keys=[1], environments="environment")
+        with self.assertRaisesRegex(exceptions.ValidationError, r".*do not allow environments.*"):
             register.RegisterService(self.mock_cp).validate_options(options)
 
     def test_does_not_allow_environment_with_consumerid(self):
@@ -571,8 +564,8 @@ class RegisterServiceTest(InjectionMockingTest):
         self.mock_cp.password = None
 
         self.mock_identity.is_valid.return_value = False
-        options = self._build_options(activation_keys=[1], consumerid='consumerid')
-        with self.assertRaisesRegex(exceptions.ValidationError, r'.*previously registered.*'):
+        options = self._build_options(activation_keys=[1], consumerid="consumerid")
+        with self.assertRaisesRegex(exceptions.ValidationError, r".*previously registered.*"):
             register.RegisterService(self.mock_cp).validate_options(options)
 
     def test_requires_basic_auth_for_normal_registration(self):
@@ -580,185 +573,6 @@ class RegisterServiceTest(InjectionMockingTest):
         self.mock_cp.password = None
 
         self.mock_identity.is_valid.return_value = False
-        options = self._build_options(consumerid='consumerid')
-        with self.assertRaisesRegex(exceptions.ValidationError, r'.*Missing username.*'):
+        options = self._build_options(consumerid="consumerid")
+        with self.assertRaisesRegex(exceptions.ValidationError, r".*Missing username.*"):
             register.RegisterService(self.mock_cp).validate_options(options)
-
-
-@subman_marker_dbus
-class DomainSocketRegisterDBusObjectTest(DBusObjectTest, InjectionMockingTest):
-    def dbus_objects(self):
-        return [RegisterDBusObject]
-
-    def setUp(self):
-        super(DomainSocketRegisterDBusObjectTest, self).setUp()
-
-        self.proxy = self.proxy_for(RegisterDBusObject.default_dbus_path)
-        self.interface = dbus.Interface(self.proxy, constants.REGISTER_INTERFACE)
-
-        self.mock_identity.is_valid.return_value = True
-
-        self.mock_cp_provider = mock.Mock(spec=CPProvider, name="CPProvider")
-        self.mock_cp = mock.Mock(spec=connection.UEPConnection, name="UEPConnection")
-
-        # Mock a basic auth connection
-        self.mock_cp.username = "username"
-        self.mock_cp.password = "password"
-
-        self.mock_cp.getOwnerList = mock.Mock()
-        self.mock_cp.getOwnerList.return_value = json.loads(OWNERS_CONTENT_JSON)
-
-        # For the tests in which it's used, the consumer_auth cp and basic_auth cp can be the same
-        self.mock_cp_provider.get_consumer_auth_cp.return_value = self.mock_cp
-        self.mock_cp_provider.get_basic_auth_cp.return_value = self.mock_cp
-
-        register_patcher = mock.patch('rhsmlib.dbus.objects.register.RegisterService', autospec=True)
-        self.mock_register = register_patcher.start().return_value
-        self.addCleanup(register_patcher.stop)
-
-        cert_invoker_patcher = mock.patch('rhsmlib.dbus.objects.register.EntCertActionInvoker', autospec=True)
-        self.mock_cert_invoker = cert_invoker_patcher.start().return_value
-        self.addCleanup(cert_invoker_patcher.stop)
-
-    def injection_definitions(self, *args, **kwargs):
-        if args[0] == inj.IDENTITY:
-            return self.mock_identity
-        elif args[0] == inj.CP_PROVIDER:
-            return self.mock_cp_provider
-        else:
-            return None
-
-    def test_open_domain_socket(self):
-        dbus_method_args = ['']
-
-        def assertions(*args):
-            result = args[0]
-            self.assertRegex(result, r'/run/dbus.*')
-
-        self.dbus_request(assertions, self.interface.Start, dbus_method_args)
-
-    def test_same_socket_on_subsequent_opens(self):
-        dbus_method_args = ['']
-
-        def assertions(*args):
-            # Assign the result as an attribute to this function.
-            # See http://stackoverflow.com/a/27910553/6124862
-            assertions.result = args[0]
-            self.assertRegex(assertions.result, r'/run/dbus.*')
-
-        self.dbus_request(assertions, self.interface.Start, dbus_method_args)
-
-        # Reset the handler_complete_event so we'll block for the second
-        # dbus_request
-        self.handler_complete_event.clear()
-
-        def assertions2(*args):
-            result2 = args[0]
-            self.assertEqual(assertions.result, result2)
-
-        self.dbus_request(assertions2, self.interface.Start, dbus_method_args)
-
-    def test_cannot_close_what_is_not_opened(self):
-        dbus_method_args = ['']
-        with self.assertRaises(dbus.exceptions.DBusException):
-            self.dbus_request(None, self.interface.Stop, dbus_method_args)
-
-    def test_closes_domain_socket(self):
-        dbus_method_args = ['']
-
-        def get_address(*args):
-            address = args[0]
-            _prefix, _equal, address = address.partition('=')
-            get_address.address, _equal, _suffix = address.partition(',')
-
-        self.dbus_request(get_address, self.interface.Start, dbus_method_args)
-        self.handler_complete_event.clear()
-
-        sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
-        try:
-            # The socket returned for connection is an abstract socket so we have
-            # to begin the name with a NUL byte to get into that namespace.  See
-            # http://blog.eduardofleury.com/archives/2007/09/13
-            sock.connect('\0' + get_address.address)
-        finally:
-            sock.close()
-
-        self.dbus_request(None, self.interface.Stop, dbus_method_args)
-        self.handler_complete_event.wait()
-
-        with self.assertRaises(socket.error) as serr:
-            sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
-            try:
-                sock.connect('\0' + get_address.address)
-            finally:
-                sock.close()
-            self.assertEqual(serr.errno, errno.ECONNREFUSED)
-
-    def _build_interface(self):
-        dbus_method_args = ['']
-
-        def get_address(*args):
-            get_address.address = args[0]
-
-        self.dbus_request(get_address, self.interface.Start, dbus_method_args)
-        self.handler_complete_event.clear()
-        socket_conn = dbus.connection.Connection(get_address.address)
-        socket_proxy = socket_conn.get_object(constants.BUS_NAME, constants.PRIVATE_REGISTER_DBUS_PATH)
-        return dbus.Interface(socket_proxy, constants.PRIVATE_REGISTER_INTERFACE)
-
-    def test_can_register_over_domain_socket(self):
-        expected_consumer = json.loads(CONSUMER_CONTENT_JSON)
-
-        def assertions(*args):
-            # Be sure we are persisting the consumer cert
-            self.assertEqual(json.loads(args[0]), expected_consumer)
-
-        self.mock_identity.is_valid.return_value = False
-        self.mock_identity.uuid = 'INVALIDCONSUMERUUID'
-
-        self.mock_register.register.return_value = expected_consumer
-
-        dbus_method_args = ['admin', 'admin', 'admin', {}, {}, '']
-        self.dbus_request(assertions, self._build_interface().Register, dbus_method_args)
-
-    def test_can_get_orgs_over_domain_socket(self):
-        expected_owners = json.loads(OWNERS_CONTENT_JSON)
-        expected_consumer = json.loads(CONSUMER_CONTENT_JSON)
-
-        def assertions(*args):
-            # Be sure we are persisting the consumer cert
-            self.assertEqual(json.loads(args[0]), expected_owners)
-
-        self.mock_identity.is_valid.return_value = False
-        self.mock_identity.uuid = 'INVALIDCONSUMERUUID'
-
-        self.mock_register.register.return_value = expected_consumer
-
-        dbus_method_args = ['admin', 'admin', {}, '']
-        self.dbus_request(assertions, self._build_interface().GetOrgs, dbus_method_args)
-
-    def test_can_register_over_domain_socket_with_activation_keys(self):
-        expected_consumer = json.loads(CONSUMER_CONTENT_JSON)
-
-        def assertions(*args):
-            # Be sure we are persisting the consumer cert
-            self.assertEqual(json.loads(args[0]), expected_consumer)
-
-        self.mock_identity.is_valid.return_value = False
-        self.mock_identity.uuid = 'INVALIDCONSUMERUUID'
-
-        self.mock_register.register.return_value = expected_consumer
-
-        dbus_method_args = [
-            'admin',
-            ['key1', 'key2'],
-            {},
-            {
-                'host': 'localhost',
-                'port': '8443',
-                'handler': '/candlepin'
-            },
-            ''
-        ]
-
-        self.dbus_request(assertions, self._build_interface().RegisterWithActivationKeys, dbus_method_args)
